Update README.md
Repository dedicated to coursework and projects for MIS-64060: Fundamentals of Machine Learning.
Project 1:
The purpose of this project was to familiarize the class with the functionality and use of R and GitHub. The main tasks for this project were to download/import a file, do general descriptive statistics, and simple graphs to review variables in the data. I use a car1 data set for my review.