Add files via upload
Repository dedicated to coursework and projects for MIS-64060: Fundamentals of Machine Learning.
Project 1:
The purpose of this project was to familiarize the class with the functionality and use of R and GitHub. The main tasks for this project was to download/import a file, do general descriptive statistics, and simple graphs to review variables in the data. I use a car1 data set for my review.
Project 2:
The purpose of this project was to further reinforce the topics of KNN classification with a data set related to bank loans. The process of determining the optimal K value was explored, as well as reviewing confusion matrices between data partitions. Additionally, predictions of specific customer profiles were reviewed.
Project 3:
The purpose of this project was to utilize the Naive Bayes algorithm to predict if flights from several different airports are likely to be delayed. This involved selecting variables of interest, converting them to factors of different levels, and pre-processing the data for the model. The results were reviewed via AUC values, ROC curves, confusion matrix and related sensitivity and specificity values for the model.
Project 4:
The purpose of this project was to utilize the K-means clustering method on university data. The data included variables on acceptance rates, enrollment rates, tuition, graduation rate, etc. I used the elbow method to determine the proper number of clusters to be used, and then the clusters were analyzed to determine the high-level groupings of these universities. Lastly, an additional record was analyzed to determine the minimum Euclidean distance between clusters to determine where it would be classified and impute missing values from that cluster.
Project 5:
The purpose of this project was to review hierarchical clustering, specifically the AGNES methodology. The different linkage methods were used (single, complete, ward, etc.) to determine the best method for a cereal review data set. Next, the appropriate number of clusters was determined via the elbow and silhouette methods. Then the stability of the clusters were reviewed by taking a partition of the data and re-assigning clusters manually. Lastly, real-world situations for normalizing and not normalizing data was discussed.
Final Exam:
The objective of this assignment was to find an optimal market segmentation for the India dish soap market, based on data collected from a marketing company. K-means clustering was utilized and analyzed for the proper number of clusters. Next, the clusters were analyzed for purchasing behaviours, brand loyalty, and demographic information. Once that information was gathered, a random forest classification model was created to target specific clusters for a promotional campaign. Findings and conclusions were to be presented in a manner appropriate for a short, concise meeting with upper management.